requirements.txt
ifcopenshell
ifctester
langgraph
langchain-core
langchain-openai
python-dotenv
